Pred zahajenim parallelizace a grafiky

+++ ppr/LISKA_PPR/PPR_Liska/PeakDetector/Peak.h
@@ -15,8 +15,8 @@ namespace PeakPeakDetector
 	{
 	public:
 		Peak(unsigned int startIndex, unsigned int endIndex,double fitness);
-		unsigned int startIndex;
-		unsigned int endIndex;
+		size_t startIndex;
+		size_t endIndex;
 		double fitness;
 		//unsigned int getStartIndex();
 		//unsigned int getEndIndex();
